Update dashboard config to match current infrastructure
`terraform plan` currently outputs many changes of the format:
`- targetAxis         = "Y1" -> null`
`- etag        = "..." -> null`
and
`- name        = "..." -> null`

The `targetAxis` field is automatically set by GCP if not set
manually by us. When this change is applied, the field remains
equal to "Y1".

Because the field is part of a json parameter, we are unable
to ignore this specific parameter.

Setting this field to "Y1" removes the diff without
changing existing infrastructure. Because we have been using
this default for some time, I don't think this is an
error in configuration.

The changes to fields "name" and "etag" are also removed with
this change. They are automatically ignored by Terraform,
but still show up in `plan` and `apply` diffs when there is
another change present.